# graph/model/models_gen.go
// Code generated by github.com/99designs/gqlgen, DO NOT EDIT.

package model

type KeyValue struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

type KeyValueInput struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

type Mutation struct {
}

type NewResourceInput struct {
	ProjectID  string           `json:"projectId"`
	LanguageID string           `json:"languageId"`
	Semver     string           `json:"semver"`
	Values     []*KeyValueInput `json:"values"`
}

type Project struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

func (Project) IsEntity() {}

type Query struct {
}

type Resource struct {
	ID         string      `json:"id"`
	ProjectID  string      `json:"projectId"`
	LanguageID string      `json:"languageId"`
	Semver     string      `json:"semver"`
	Values     []*KeyValue `json:"values"`
}

func (Resource) IsEntity() {}